Lots of clean up
Please check that my cleaning didn't break your current correction.
